Implement unified Material Design 3 color system and migration guide
- Added `migration_guide.rs` to assist in transitioning from dual MaterialColors to a unified system.
- Introduced `unified_colors.rs` for the complete implementation of Material Design 3 colors with field-based access.
- Updated `mod.rs` to deprecate old color modules and re-export unified color types.
- Refactored `core.rs` to utilize the new unified color system.
- Modified `theme.rs` to use the new MaterialColors structure.
- Enhanced `settings.rs` to integrate Material Design 3 selection components and improve styling.
- Added compatibility methods in `MaterialColors` for legacy support.

+++ abop-gui/src/styling/material/components/selection/chip/core.rs
@@ -17,7 +17,7 @@ use std::time::Duration;
 pub const MAX_CHIP_LABEL_LENGTH: usize = constants::chips::MAX_LABEL_LENGTH;
 
 /// Default animation duration for chip state transitions
-pub const DEFAULT_ANIMATION_DURATION: Duration = Duration::from_millis(constants::animation::FAST_DURATION_MS);
+pub const DEFAULT_ANIMATION_DURATION: Duration = Duration::from_millis(constants::animation::FAST_DURATION_MS as u64);
 
 // ============================================================================
 // Core Chip Implementation Extensions
